requirements.txt
Pillow
matplotlib
opencv-python
torch
